% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline.R
\name{ridgeline}
\alias{ridgeline}
\title{Ridgeline plot}
\usage{
ridgeline(
  x,
  y,
  bw = "nrd0",
  mode = FALSE,
  main,
  xlab,
  ylab,
  col = "gray",
  border,
  lty = 1,
  lwd = 1,
  bty = "o",
  labels = NULL,
  palette,
  axes = TRUE
)
}
\arguments{
\item{x}{Numeric variable.}

\item{y}{Categorical variable.}

\item{bw}{Bandwidth value or method. See \code{?density} for more details.}

\item{mode}{Whether to display the modes or not}

\item{main}{Title of the plot}

\item{xlab}{X-axis title of the plot}

\item{ylab}{Y-axis title of the plot}

\item{col}{Color of the horizontal lines for each group}

\item{border}{Color of the borders of the densities}

\item{lty}{Line type of the border of the densities}

\item{lwd}{Line width of the border of the densities}

\item{bty}{Box type of the plot}

\item{labels}{If specified, overrides the labels of the Y-axis}

\item{palette}{Color, colors or a color palette for the density areas}

\item{axes}{If FALSE, removes the X-axis}
}
\description{
Ridgeline plot
}
